server/modal/requirements.txt
openai
fastapi
pydantic
python-dotenv
uvicorn
